moving TestHelpers to test.rs and renaming native.rs to llvm.rs
